Updated Comments - getRandom Functions
I had to update the comments around the getRandomColor and
getRandomQuote functions as I had originally stated that these would
make it so that the colors and quotes would never repeat. This isn’t
true though as it’s possible that the last spliced index value of each
respective array might be first once it’s refilled and the function is
called again.